services/chatbot-service/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3
mysqlclient==2.2.0
redis==4.5.4
celery==5.3.1
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
openai==1.3.0
python-decouple==3.8
requests==2.31.0
Pillow==10.0.1
django-extensions==3.2.3
PyJWT==2.8.0
cryptography==41.0.7
markdown==3.5.1
scikit-learn==1.3.2
numpy==1.24.3
nltk==3.8.1
fuzzywuzzy==0.18.0
python-Levenshtein==0.21.1
beautifulsoup4==4.12.2
lxml==4.9.3
jsonschema==4.19.2
python-dotenv==1.0.0
